requirements.txt
defusedxml==0.4.1
Django==1.7.5
django-admin-bootstrapped==2.3.5
django-guardian==1.2.5
django-staticfiles-bootstrap==3.0.3
django-widget-tweaks==1.3
geopy==1.9.1
hashids==1.0.3
mysqlclient==1.3.6
oauthlib==0.7.2
pygeoip==0.3.2
PyJWT==0.4.1
python-social-auth==0.2.2
python3-openid==3.0.5
pytz==2014.10
requests==2.5.3
requests-oauthlib==0.3.1
six==1.9.0
django-debug-toolbar
django-debug-panel
django-mathfilters